+++ plotting/current_annealing/__init__.py
@@ -1,2 +1,0 @@
-"""Current annealing plotting package."""
-